LayoutTests/platform/mac-elcapitan/fast/text/font-selection-font-loading-api-parse-expected.txt

FAIL setFontFaceWeight(fontFace, '1') should be 1. Threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern.
PASS setFontFaceWeight(fontFace, '100') is "100"
PASS setFontFaceWeight(fontFace, '200') is "200"
PASS setFontFaceWeight(fontFace, '300') is "300"
PASS setFontFaceWeight(fontFace, '400') is "normal"
PASS setFontFaceWeight(fontFace, 'normal') is "normal"
PASS setFontFaceWeight(fontFace, '500') is "500"
PASS setFontFaceWeight(fontFace, '600') is "600"
PASS setFontFaceWeight(fontFace, '700') is "bold"
PASS setFontFaceWeight(fontFace, 'bold') is "bold"
PASS setFontFaceWeight(fontFace, '800') is "800"
PASS setFontFaceWeight(fontFace, '900') is "900"
PASS setFontFaceWeight(fontFace, 'garbage') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS setFontFaceWeight(fontFace, 'initial') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
FAIL setFontFaceWeight(fontFace, 'calc(3 + 4)') should be 7. Threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern.
PASS setFontFaceWeight(fontFace, 'calc(100 + 200)') is "300"
PASS setFontFaceWeight(fontFace, 'calc(150 + 50)') is "200"
FAIL setFontFaceStretch(fontFace, '1%') should be 1%. Threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern.
PASS setFontFaceStretch(fontFace, '2') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS setFontFaceStretch(fontFace, 'ultra-condensed') is "ultra-condensed"
PASS setFontFaceStretch(fontFace, 'extra-condensed') is "extra-condensed"
PASS setFontFaceStretch(fontFace, 'condensed') is "condensed"
PASS setFontFaceStretch(fontFace, 'semi-condensed') is "semi-condensed"
PASS setFontFaceStretch(fontFace, 'normal') is "normal"
PASS setFontFaceStretch(fontFace, 'semi-expanded') is "semi-expanded"
PASS setFontFaceStretch(fontFace, 'expanded') is "expanded"
PASS setFontFaceStretch(fontFace, 'extra-expanded') is "extra-expanded"
PASS setFontFaceStretch(fontFace, 'ultra-expanded') is "ultra-expanded"
PASS setFontFaceStretch(fontFace, 'garbage') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS setFontFaceStretch(fontFace, 'initial') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS setFontFaceStretch(fontFace, 'calc(3 + 4)') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
FAIL setFontFaceStretch(fontFace, 'calc(3% + 4%)') should be 7%. Threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern.
FAIL setFontFaceStyle(fontFace, 'oblique 1deg') should be oblique 1deg. Threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern.
FAIL setFontFaceStyle(fontFace, 'oblique 20grad') should be oblique 18deg. Threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern.
FAIL setFontFaceStyle(fontFace, 'oblique 0.28318rad') should be oblique 16deg. Threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern.
FAIL setFontFaceStyle(fontFace, 'oblique 0.04turn') should be oblique 14.25deg. Threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern.
PASS setFontFaceStyle(fontFace, 'oblique 5') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS setFontFaceStyle(fontFace, 'oblique 20') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS setFontFaceStyle(fontFace, 'italic') is "italic"
PASS setFontFaceStyle(fontFace, 'oblique') is "italic"
PASS setFontFaceStyle(fontFace, 'normal') is "normal"
PASS setFontFaceStyle(fontFace, 'garbage') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS setFontFaceStyle(fontFace, 'initial') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
PASS setFontFaceStyle(fontFace, 'oblique calc(3 + 4)') threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern..
FAIL setFontFaceStyle(fontFace, 'oblique calc(0.04turn + 1deg)') should be oblique 15.25deg. Threw exception SyntaxError (DOM Exception 12): The string did not match the expected pattern.
PASS successfullyParsed is true

TEST COMPLETE
